+++ CHANGELOG.md
@@ -10,6 +10,8 @@
 - Implemented pipeline status on status bar.
 - Implemented open active file on GitLab including active line number and selection.
 - Implemented open current MR on GitLab.
+- Implemented open GitLab to create new merge request.
+- Implemented open GitLab to create new issue.
 
 ### Changed
 - Removed `gitlab.userId` necessity.